adding SpliceWord and refactoring code
