+++ cpp/src/arrow/util/CMakeLists.txt
@@ -44,7 +44,6 @@ add_arrow_test(utility-test
                async_generator_test.cc
                bit_block_counter_test.cc
                bit_util_test.cc
-               bitmap_test.cc
                cache_test.cc
                checked_cast_test.cc
                compression_test.cc
